# openapi.json
{"openapi":"3.1.0","info":{"title":"Sonarr API Tool Server","description":"Enhanced OpenAPI server for Sonarr integration with Open WebUI - supports full series, episode, and file management","version":"2.0.0"},"paths":{"/":{"get":{"summary":"Health check","description":"Basic health check endpoint","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/sonarr/search":{"get":{"summary":"Search for TV series","description":"Search for TV series in Sonarr","operationId":"search_series_sonarr_search_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"term","in":"query","required":true,"schema":{"type":"string","description":"Search term for TV series","title":"Term"},"description":"Search term for TV series"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SeriesSearchResult"},"title":"Response Search Series Sonarr Search Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/sonarr/series":{"get":{"summary":"List all series","description":"Get all series with their details","operationId":"list_series_sonarr_series_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/SeriesDetail"},"type":"array","title":"Response List Series Sonarr Series Get"}}}}},"security":[{"HTTPBearer":[]}]},"post":{"summary":"Add a TV series to Sonarr","description":"Add a TV series to Sonarr with specified configuration","operationId":"add_series_sonarr_series_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddSeriesRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SeriesDetail"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/sonarr/series/{series_id}":{"get":{"summary":"Get specific series details","description":"Get detailed information about a specific series including tags, quality profile, and root folder","operationId":"get_series_sonarr_series__series_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"series_id","in":"path","required":true,"schema":{"type":"integer","title":"Series Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SeriesDetail"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"summary":"Update series settings","description":"Update series settings including root folder, quality profile, monitoring, and tags","operationId":"update_series_sonarr_series__series_id__put","security":[{"HTTPBearer":[]}],"parameters":[{"name":"series_id","in":"path","required":true,"schema":{"type":"integer","title":"Series Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateSeriesRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SeriesDetail"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/sonarr/episodes":{"get":{"summary":"Get episodes for a series","description":"Get all episodes for a series","operationId":"get_episodes_sonarr_episodes_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"series_id","in":"query","required":true,"schema":{"type":"integer","title":"Series Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/EpisodeDetail"},"title":"Response Get Episodes Sonarr Episodes Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/sonarr/episodes/{episode_id}":{"put":{"summary":"Update episode monitoring","description":"Update monitoring status for a single episode","operationId":"update_episode_sonarr_episodes__episode_id__put","security":[{"HTTPBearer":[]}],"parameters":[{"name":"episode_id","in":"path","required":true,"schema":{"type":"integer","title":"Episode Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"boolean","title":"Monitored"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/sonarr/episodes/monitor":{"put":{"summary":"Bulk update episode monitoring","description":"Update monitoring status for multiple episodes","operationId":"bulk_update_episodes_sonarr_episodes_monitor_put","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BulkEpisodeUpdate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/sonarr/series/{series_id}/season":{"put":{"summary":"Update season monitoring","description":"Update monitoring for an entire season","operationId":"update_season_monitoring_sonarr_series__series_id__season_put","security":[{"HTTPBearer":[]}],"parameters":[{"name":"series_id","in":"path","required":true,"schema":{"type":"integer","title":"Series Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SeasonMonitorRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/sonarr/wanted/missing":{"get":{"summary":"Get missing episodes for a specific series","description":"Get list of missing episodes for a specific series","operationId":"get_missing_episodes_by_series_sonarr_wanted_missing_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"seriesId","in":"query","required":true,"schema":{"type":"integer","description":"Filter by series ID","title":"Seriesid"},"description":"Filter by series ID"},{"name":"page","in":"query","required":false,"schema":{"type":"integer","minimum":1,"default":1,"title":"Page"}},{"name":"pageSize","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":50,"title":"Pagesize"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MissingEpisode"},"title":"Response Get Missing Episodes By Series Sonarr Wanted Missing Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/sonarr/wanted/missing/all":{"get":{"summary":"Get all missing episodes","description":"Get list of all missing episodes","operationId":"get_all_missing_episodes_sonarr_wanted_missing_all_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","minimum":1,"default":1,"title":"Page"}},{"name":"pageSize","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":50,"title":"Pagesize"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MissingEpisode"},"title":"Response Get All Missing Episodes Sonarr Wanted Missing All Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/sonarr/wanted/missing/search":{"post":{"summary":"Search for missing episodes","description":"Trigger search for missing episodes","operationId":"search_missing_episodes_sonarr_wanted_missing_search_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"seriesId","in":"query","required":false,"schema":{"type":"integer","description":"Series ID to search (0 for all)","default":0,"title":"Seriesid"},"description":"Series ID to search (0 for all)"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/sonarr/tags":{"get":{"summary":"List all tags","description":"Get all tags","operationId":"list_tags_sonarr_tags_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/TagModel"},"type":"array","title":"Response List Tags Sonarr Tags Get"}}}}},"security":[{"HTTPBearer":[]}]},"post":{"summary":"Create a new tag","description":"Create a new tag","operationId":"create_tag_sonarr_tags_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TagModel"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TagModel"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/sonarr/qualityprofiles":{"get":{"summary":"List quality profiles","description":"Get all quality profiles","operationId":"list_quality_profiles_sonarr_qualityprofiles_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/sonarr/rootfolders":{"get":{"summary":"List root folders","description":"Get all root folders","operationId":"list_root_folders_sonarr_rootfolders_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/sonarr/command":{"post":{"summary":"Execute Sonarr commands","description":"Execute various Sonarr commands like RefreshSeries, RssSync, etc.","operationId":"execute_command_sonarr_command_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommandRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}}},"components":{"schemas":{"AddSeriesRequest":{"properties":{"tvdbId":{"type":"integer","title":"Tvdbid","description":"The TVDB ID of the series to add"},"qualityProfileId":{"type":"integer","title":"Qualityprofileid","description":"Quality profile ID","default":0},"rootFolderPath":{"type":"string","title":"Rootfolderpath","description":"Root folder path","default":""},"monitored":{"type":"boolean","title":"Monitored","description":"Whether to monitor the series","default":true},"seasonFolder":{"type":"boolean","title":"Seasonfolder","description":"Whether to use season folders","default":true},"tags":{"items":{"type":"integer"},"type":"array","title":"Tags","description":"Tag IDs to apply","default":[]},"addOptions":{"additionalProperties":true,"type":"object","title":"Addoptions","default":{"monitor":"all","searchForMissingEpisodes":true,"searchForCutoffUnmetEpisodes":false}}},"type":"object","required":["tvdbId"],"title":"AddSeriesRequest"},"BulkEpisodeUpdate":{"properties":{"episodeIds":{"items":{"type":"integer"},"type":"array","title":"Episodeids"},"monitored":{"type":"boolean","title":"Monitored"}},"type":"object","required":["episodeIds","monitored"],"title":"BulkEpisodeUpdate"},"CommandRequest":{"properties":{"name":{"type":"string","title":"Name"},"seriesId":{"type":"integer","title":"Seriesid","description":"Filter by series ID","default":0},"episodeIds":{"items":{"type":"integer"},"type":"array","title":"Episodeids","description":"Episode IDs"}},"type":"object","required":["name"],"title":"CommandRequest"},"EpisodeDetail":{"properties":{"id":{"type":"integer","title":"Id"},"seriesId":{"type":"integer","title":"Seriesid"},"seasonNumber":{"type":"integer","title":"Seasonnumber"},"episodeNumber":{"type":"integer","title":"Episodenumber"},"title":{"type":"string","title":"Title"},"airDate":{"type":"string","title":"Airdate","default":""},"monitored":{"type":"boolean","title":"Monitored"},"hasFile":{"type":"boolean","title":"Hasfile"},"episodeFileId":{"type":"integer","title":"Episodefileid","default":0}},"type":"object","required":["id","seriesId","seasonNumber","episodeNumber","title","monitored","hasFile"],"title":"EpisodeDetail"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"MissingEpisode":{"properties":{"id":{"type":"integer","title":"Id"},"seriesTitle":{"type":"string","title":"Seriestitle"},"seasonNumber":{"type":"integer","title":"Seasonnumber"},"episodeNumber":{"type":"integer","title":"Episodenumber"},"title":{"type":"string","title":"Title"},"airDate":{"type":"string","title":"Airdate","default":""}},"type":"object","required":["id","seriesTitle","seasonNumber","episodeNumber","title"],"title":"MissingEpisode"},"SeasonMonitorRequest":{"properties":{"seasonNumber":{"type":"integer","title":"Seasonnumber"},"monitored":{"type":"boolean","title":"Monitored"}},"type":"object","required":["seasonNumber","monitored"],"title":"SeasonMonitorRequest"},"SeriesDetail":{"properties":{"id":{"type":"integer","title":"Id","default":0},"title":{"type":"string","title":"Title"},"tvdbId":{"type":"integer","title":"Tvdbid"},"path":{"type":"string","title":"Path"},"rootFolderPath":{"type":"string","title":"Rootfolderpath"},"qualityProfileId":{"type":"integer","title":"Qualityprofileid"},"qualityProfileName":{"type":"string","title":"Qualityprofilename","default":""},"monitored":{"type":"boolean","title":"Monitored"},"seasonFolder":{"type":"boolean","title":"Seasonfolder"},"tags":{"items":{"type":"integer"},"type":"array","title":"Tags","default":[]},"tagNames":{"items":{"type":"string"},"type":"array","title":"Tagnames","default":[]},"statistics":{"additionalProperties":true,"type":"object","title":"Statistics","default":{}},"seasons":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Seasons","default":[]}},"type":"object","required":["title","tvdbId","path","rootFolderPath","qualityProfileId","monitored","seasonFolder"],"title":"SeriesDetail"},"SeriesSearchResult":{"properties":{"title":{"type":"string","title":"Title"},"tvdbId":{"type":"integer","title":"Tvdbid"},"year":{"type":"integer","title":"Year","default":0},"overview":{"type":"string","title":"Overview","default":""},"remotePoster":{"type":"string","title":"Remoteposter","default":""},"added":{"type":"boolean","title":"Added","default":false}},"type":"object","required":["title","tvdbId"],"title":"SeriesSearchResult"},"TagModel":{"properties":{"id":{"type":"integer","title":"Id","default":0},"label":{"type":"string","title":"Label"}},"type":"object","required":["label"],"title":"TagModel"},"UpdateSeriesRequest":{"properties":{"rootFolderPath":{"type":"string","title":"Rootfolderpath","default":""},"qualityProfileId":{"type":"integer","title":"Qualityprofileid","default":0},"monitored":{"type":"boolean","title":"Monitored","default":false},"seasonFolder":{"type":"boolean","title":"Seasonfolder","default":false},"tags":{"items":{"type":"integer"},"type":"array","title":"Tags","default":[]},"moveFiles":{"type":"boolean","title":"Movefiles","default":false}},"type":"object","title":"UpdateSeriesRequest"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}